+++ Realm/Realm/Handles/RealmHandle.cs
@@ -148,7 +148,7 @@ namespace Realms
             }
             catch(Exception ex)
             {
-                Logger.Default.Log(LogLevel.Error, $"An error occurred while closing native handle. Please file an issue at https://github.com/realm/realm-dotnet/issues. Error: {ex}");
+                RealmLogger.Default.Log(LogLevel.Error, $"An error occurred while closing native handle. Please file an issue at https://github.com/realm/realm-dotnet/issues. Error: {ex}");
                 Debug.Fail($"Failed to close native handle: {ex}");
 
                 // it would be really bad if we got an exception in here. We must not pass it on, but have to return false