+++ Realm/Realm/Logging/Logger.cs
@@ -25,25 +25,43 @@ using System.Threading.Tasks;
 
 namespace Realms.Logging
 {
+    /// <inheritdoc/>
+    [Obsolete("Use RealmLogger instead. If using a custom logger, RealmLogger.LogImpl() additionally receives the log category.")]
+    public abstract class Logger : RealmLogger
+    {
+        /// <summary>
+        /// The internal implementation being called from <see cref="RealmLogger.Log(LogLevel, string)"/>.
+        /// </summary>
+        /// <param name="level">The criticality level for the message.</param>
+        /// <param name="message">The message to log.</param>
+        protected abstract void LogImpl(LogLevel level, string message);
+
+        /// <inheritdoc/>
+        protected override void LogImpl(LogLevel level, LogCategory category, string message)
+        {
+            LogImpl(level, message);
+        }
+    }
+
     /// <summary>
     /// A logger that logs messages originating from Realm. The default logger can be replaced by setting <see cref="Default"/>.
     /// <br/>
     /// A few built-in implementations are provided by <see cref="Console"/>, <see cref="Null"/>, and <see cref="Function(Action{string})"/>,
     /// but you can implement your own.
     /// </summary>
-    public abstract class Logger
+    public abstract class RealmLogger
     {
         private readonly Lazy<GCHandle> _gcHandle;
 
-        private static Logger? _defaultLogger;
-        private static LogLevel _logLevel = LogLevel.Info;
+        private static readonly LogCategory _defaultLogCategory = LogCategory.Realm;
+        private static RealmLogger? _defaultLogger;
 
         /// <summary>
         /// Gets a <see cref="ConsoleLogger"/> that outputs messages to the default console. For most project types, that will be
         /// using <see cref="Console.WriteLine()"/> but certain platforms may use different implementations.
         /// </summary>
-        /// <value>A <see cref="Logger"/> instance that outputs to the platform's console.</value>
-        public static Logger Console { get; internal set; } = new ConsoleLogger();
+        /// <value>A <see cref="RealmLogger"/> instance that outputs to the platform's console.</value>
+        public static RealmLogger Console { get; internal set; } = new ConsoleLogger();
 
         /// <summary>
         /// Gets a <see cref="FileLogger"/> that saves the log messages to a file.
@@ -54,24 +72,15 @@ namespace Realms.Logging
         /// Please note that this logger is not optimized for performance, and could lead to overall sync performance slowdown with more verbose log levels.
         /// </remarks>
         /// <returns>
-        /// A <see cref="Logger"/> instance that will save log messages to a file.
+        /// A <see cref="RealmLogger"/> instance that will save log messages to a file.
         /// </returns>
-        public static Logger File(string filePath, Encoding? encoding = null) => new FileLogger(filePath, encoding);
+        public static RealmLogger File(string filePath, Encoding? encoding = null) => new FileLogger(filePath, encoding);
 
         /// <summary>
         /// Gets a <see cref="NullLogger"/> that ignores all messages.
         /// </summary>
-        /// <value>A <see cref="Logger"/> that doesn't output any messages.</value>
-        public static Logger Null { get; } = new NullLogger();
-
-        /// <summary>
-        /// Gets a <see cref="FunctionLogger"/> that proxies Log calls to the supplied function.
-        /// </summary>
-        /// <param name="logFunction">Function to proxy log calls to.</param>
-        /// <returns>
-        /// A <see cref="Logger"/> instance that will invoke <paramref name="logFunction"/> for each message.
-        /// </returns>
-        public static Logger Function(Action<LogLevel, string> logFunction) => new FunctionLogger(logFunction);
+        /// <value>A <see cref="RealmLogger"/> that doesn't output any messages.</value>
+        public static RealmLogger Null { get; } = new NullLogger();
 
         /// <summary>
         /// Gets a <see cref="FunctionLogger"/> that proxies Log calls to the supplied function. The message will
@@ -79,30 +88,73 @@ namespace Realms.Logging
         /// </summary>
         /// <param name="logFunction">Function to proxy log calls to.</param>
         /// <returns>
-        /// A <see cref="Logger"/> instance that will invoke <paramref name="logFunction"/> for each message.
+        /// A <see cref="RealmLogger"/> instance that will invoke <paramref name="logFunction"/> for each message.
         /// </returns>
-        public static Logger Function(Action<string> logFunction) => new FunctionLogger((level, message) => logFunction(FormatLog(level, message)));
+        public static RealmLogger Function(Action<string> logFunction) => new FunctionLogger((level, category, message) => logFunction(FormatLog(level, category, message)));
 
         /// <summary>
-        /// Gets or sets the verbosity of log messages.
+        /// Gets a <see cref="FunctionLogger"/> that proxies Log calls to the supplied function.
+        /// </summary>
+        /// <param name="logFunction">Function to proxy log calls to.</param>
+        /// <returns>
+        /// A <see cref="RealmLogger"/> instance that will invoke <paramref name="logFunction"/> for each message.
+        /// </returns>
+        [Obsolete("Use Function(Action<LogLevel, LogCategory, string> logFunction).")]
+        public static RealmLogger Function(Action<LogLevel, string> logFunction) => new FunctionLogger((level, _, message) => logFunction(level, message));
+
+        /// <summary>
+        /// Gets a <see cref="FunctionLogger"/> that proxies Log calls to the supplied function.
+        /// </summary>
+        /// <param name="logFunction">Function to proxy log calls to.</param>
+        /// <returns>
+        /// A <see cref="RealmLogger"/> instance that will invoke <paramref name="logFunction"/> for each message.
+        /// </returns>
+        public static RealmLogger Function(Action<LogLevel, LogCategory, string> logFunction) => new FunctionLogger(logFunction);
+
+        /// <summary>
+        /// Gets or sets the verbosity of log messages for all log categories via <see cref="LogCategory.Realm"/>.
         /// </summary>
         /// <value>The log level for Realm-originating messages.</value>
+        [Obsolete("Use GetLogLevel() and SetLogLevel().")]
         public static LogLevel LogLevel
         {
-            get => _logLevel;
+            get => GetLogLevel();
             set
             {
-                _logLevel = value;
-                SharedRealmHandle.SetLogLevel(value);
+                SetLogLevel(value);
             }
         }
 
         /// <summary>
-        /// Gets or sets a custom <see cref="Logger"/> implementation that will be used by
+        /// Gets the verbosity of log messages for the given category.
+        /// </summary>
+        /// <param name="category">The category to get the level for. Defaults to <see cref="LogCategory.Realm"/> if not specified.</param>
+        /// <returns>
+        /// The log level used for the given category.
+        /// </returns>
+        public static LogLevel GetLogLevel(LogCategory? category = null)
+        {
+            category ??= _defaultLogCategory;
+            return SharedRealmHandle.GetLogLevel(category);
+        }
+
+        /// <summary>
+        /// Sets the verbosity of log messages for the given category.
+        /// </summary>
+        /// <param name="level">The log level to use for messages.</param>
+        /// <param name="category">The category to set the level for. Defaults to <see cref="LogCategory.Realm"/> if not specified.</param>
+        public static void SetLogLevel(LogLevel level, LogCategory? category = null)
+        {
+            category ??= _defaultLogCategory;
+            SharedRealmHandle.SetLogLevel(level, category);
+        }
+
+        /// <summary>
+        /// Gets or sets a custom <see cref="RealmLogger"/> implementation that will be used by
         /// Realm whenever information must be logged.
         /// </summary>
         /// <value>The logger to be used for Realm-originating messages.</value>
-        public static Logger Default
+        public static RealmLogger Default
         {
             get => _defaultLogger ?? Console;
             set => _defaultLogger = value;
@@ -111,57 +163,76 @@ namespace Realms.Logging
         internal GCHandle GCHandle => _gcHandle.Value;
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="Logger"/> class.
+        /// Initializes a new instance of the <see cref="RealmLogger"/> class.
         /// </summary>
-        protected Logger()
+        protected RealmLogger()
         {
             _gcHandle = new Lazy<GCHandle>(() => GCHandle.Alloc(this));
         }
 
-        internal static void LogDefault(LogLevel level, string message) => Default?.Log(level, message);
-
         /// <summary>
-        /// Log a message at the supplied level.
+        /// Log a message at the supplied level and default category <see cref="LogCategory.RealmLogCategory.SDK"/>.
         /// </summary>
         /// <param name="level">The criticality level for the message.</param>
         /// <param name="message">The message to log.</param>
         public void Log(LogLevel level, string message)
         {
-            if (level < LogLevel)
+            Log(level, LogCategory.Realm.SDK, message);
+        }
+
+        /// <summary>
+        /// Log a message at the supplied level and category.
+        /// </summary>
+        /// <param name="level">The criticality level for the message.</param>
+        /// <param name="category">The category for the message.</param>
+        /// <param name="message">The message to log.</param>
+        public void Log(LogLevel level, LogCategory category, string message)
+        {
+            if (level < GetLogLevel(category))
             {
                 return;
             }
 
+            LogAnyLevel(level, category, message);
+        }
+
+        /// <summary>
+        /// Log a message without calling into Core to check the current level. Logs from
+        /// Core should always call this API as they already check the level prior to notifying.
+        /// </summary>
+        internal void LogAnyLevel(LogLevel level, LogCategory category, string message)
+        {
             try
             {
-                LogImpl(level, message);
+                LogImpl(level, category, message);
             }
             catch (Exception ex)
             {
-                Console.Log(LogLevel.Error, $"An exception occurred while trying to log the message: '{message}' at level: {level}. Error: {ex}");
+                Console.Log(LogLevel.Error, $"An exception occurred while trying to log the message: '{message}' at level: '{level}' in category: '{category}'. Error: {ex}");
             }
         }
 
         /// <summary>
-        /// The internal implementation being called from <see cref="Log"/>.
+        /// The internal implementation being called from <see cref="Log(Realms.Logging.LogLevel, LogCategory, string)"/>.
         /// </summary>
         /// <param name="level">The criticality level for the message.</param>
+        /// <param name="category">The category for the message.</param>
         /// <param name="message">The message to log.</param>
-        protected abstract void LogImpl(LogLevel level, string message);
+        protected abstract void LogImpl(LogLevel level, LogCategory category, string message);
 
-        internal static string FormatLog(LogLevel level, string message) => $"{DateTimeOffset.UtcNow:yyyy-MM-dd HH:mm:ss.fff} {level}: {message}";
+        internal static string FormatLog(LogLevel level, LogCategory category, string message) => $"{DateTimeOffset.UtcNow:yyyy-MM-dd HH:mm:ss.fff} {category} {level}: {message}";
 
-        private class ConsoleLogger : Logger
+        private class ConsoleLogger : RealmLogger
         {
-            protected override void LogImpl(LogLevel level, string message)
+            protected override void LogImpl(LogLevel level, LogCategory category, string message)
             {
-                System.Console.WriteLine(FormatLog(level, message));
+                System.Console.WriteLine(FormatLog(level, category, message));
             }
         }
 
-        private class FileLogger : Logger
+        private class FileLogger : RealmLogger
         {
-            private readonly object locker = new();
+            private readonly object _locker = new();
             private readonly string _filePath;
             private readonly Encoding _encoding;
 
@@ -171,43 +242,43 @@ namespace Realms.Logging
                 _encoding = encoding ?? Encoding.UTF8;
             }
 
-            protected override void LogImpl(LogLevel level, string message)
+            protected override void LogImpl(LogLevel level, LogCategory category, string message)
             {
-                lock (locker)
+                lock (_locker)
                 {
-                    System.IO.File.AppendAllText(_filePath, FormatLog(level, message) + Environment.NewLine, _encoding);
+                    System.IO.File.AppendAllText(_filePath, FormatLog(level, category, message) + Environment.NewLine, _encoding);
                 }
             }
         }
 
-        private class FunctionLogger : Logger
+        private class FunctionLogger : RealmLogger
         {
-            private readonly Action<LogLevel, string> _logFunction;
+            private readonly Action<LogLevel, LogCategory, string> _logFunction;
 
-            public FunctionLogger(Action<LogLevel, string> logFunction)
+            public FunctionLogger(Action<LogLevel, LogCategory, string> logFunction)
             {
                 _logFunction = logFunction;
             }
 
-            protected override void LogImpl(LogLevel level, string message) => _logFunction(level, message);
+            protected override void LogImpl(LogLevel level, LogCategory category, string message) => _logFunction(level, category, message);
         }
 
-        private class NullLogger : Logger
+        private class NullLogger : RealmLogger
         {
-            protected override void LogImpl(LogLevel level, string message)
+            protected override void LogImpl(LogLevel level, LogCategory category, string message)
             {
             }
         }
 
-        internal class InMemoryLogger : Logger
+        internal class InMemoryLogger : RealmLogger
         {
             private readonly StringBuilder _builder = new();
 
-            protected override void LogImpl(LogLevel level, string message)
+            protected override void LogImpl(LogLevel level, LogCategory category, string message)
             {
                 lock (_builder)
                 {
-                    _builder.AppendLine(FormatLog(level, message));
+                    _builder.AppendLine(FormatLog(level, category, message));
                 }
             }
 
@@ -228,7 +299,7 @@ namespace Realms.Logging
             }
         }
 
-        internal class AsyncFileLogger : Logger, IDisposable
+        internal class AsyncFileLogger : RealmLogger, IDisposable
         {
             private readonly ConcurrentQueue<string> _queue = new();
             private readonly string _filePath;
@@ -255,11 +326,11 @@ namespace Realms.Logging
                 _flush.Dispose();
             }
 
-            protected override void LogImpl(LogLevel level, string message)
+            protected override void LogImpl(LogLevel level, LogCategory category, string message)
             {
                 if (!_isFlushing)
                 {
-                    _queue.Enqueue(FormatLog(level, message));
+                    _queue.Enqueue(FormatLog(level, category, message));
                     _hasNewItems.Set();
                 }
             }